Task related classes moved under Task folder
