+++ model/WorkerInterface.php
@@ -20,6 +20,7 @@
 
 namespace oat\taoTaskQueue\model;
 
+use oat\taoTaskQueue\model\Task\TaskInterface;
 use Psr\Log\LoggerAwareInterface;
 
 /**
